Add files via upload
This data visualization was created for my final presentation in BIO 485, Principles of Bioinformatics. My partner and I chose the 3rd project option, differential gene expression analysis by RNA-Seq. We performed differential gene expression analysis of Mus musculus using both RSEM and Kallisto-DESeq2, focusing on genes involved in eye development. This required us to use Miami’s supercomputer cluster. Additionally, I conducted some data cleaning and analysis in R, which I also used to produce several data visualizations. I chose to take the log of the transcripts per million in order to make it easier for our classmates to visualize the differences in gene expression between the two cell types, Epithelial and Fiber, as well as the two differential gene expression tools, Kallisto and RSEM. Although a red and blue color palette is traditional, I chose to use different shades of red, yellow, and blue in order to make the visualization more accessible to color-blind classmates. As part of this project, in addition to presenting our results to the class, we also produced a literature review.